requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# LLM Clients
openai==1.12.0
anthropic==0.18.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Database
sqlalchemy==2.0.25

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Code Quality
black==24.1.1
flake8==7.0.0